feat(feishu-doc): add atomic create_and_write flow
